backend/requirements.txt
# FastAPI and ASGI
fastapi==0.135.3
uvicorn[standard]==0.42.0

# Database
sqlalchemy[asyncio]==2.0.49
asyncpg==0.29.0
alembic==1.13.1

# Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.12

# Configuration
pydantic-settings==2.7.1
python-dotenv==1.2.2

# Email
aiosmtplib==3.0.2

# Document Processing
reportlab==4.4.10
pymupdf==1.24.1
python-docx==1.2.0
pandas==2.2.2
openpyxl==3.1.2
# jinja2 for email templates
jinja2==3.1.4

# AI Paraphrasing
google-genai==0.3.0
httpx==0.27.0
